+++ addon/mixins/select-box/select-active-option-on-enter.js
@@ -1,7 +1,7 @@
 import Mixin from '@ember/object/mixin';
 
 export default Mixin.create({
-  pressEnter(...args) {
+  _onPressEnter(...args) {
     this._super(...args);
     this.send('selectActiveOption');
   }