+++ addon/utils/invoke-action.js
@@ -1,7 +1,7 @@
 export default function invokeAction(object, name, ...args) {
-  const action = object[name];
+  const func = object[name];
 
-  if (typeof action === 'function') {
-    return action(...args);
+  if (typeof func === 'function') {
+    return func.apply(object, args);
   }
 }